Move swap navigation state and quote inputs to callers
