+++ src/__swaps__/screens/Swap/components/TokenList/constants.ts
@@ -1,0 +1,2 @@
+export const BUY_LIST_HEADER_HEIGHT = 20 + 10 + 10;
+export const SELL_LIST_HEADER_HEIGHT = 20 + 10 + 14;
