+++ src/__swaps__/screens/Swap/hooks/useBottomPanelGestureHandler.ts
@@ -1,7 +1,8 @@
 import { Gesture } from 'react-native-gesture-handler';
 import { interpolate, useSharedValue, withSpring } from 'react-native-reanimated';
 
-import { NavigationSteps, useSwapContext } from '@/__swaps__/screens/Swap/providers/swap-provider';
+import { NavigationSteps } from '@/__swaps__/screens/Swap/hooks/useSwapNavigation';
+import { useSwapContext } from '@/__swaps__/screens/Swap/providers/swap-provider';
 import { SPRING_CONFIGS } from '@/components/animations/animationConfigs';
 
 export const useBottomPanelGestureHandler = () => {
